Add `get_inputs_generator` classmethod to `CommonRelaxWorkChain`
This method serves as a proxy to generate an instance of the
`InputsGenerator` that is implemented for the subclass of the common
relax work chain. The work chain sub class should just define the
`_generator_class` class attribute to the subclass of the
`RelaxInputsGenerator`, just like it defines the `_process_class`.

Since the work chain will be the main point of entry for all
interactions, this way any client can obtain an instance of the input
generator which in turn provides all the information about available
protocols and the actual method to generate the inputs.